cleaned up project to be packet only and made server more structured
