+++ ArgoProject/ArgoProject.vcxproj.filters
@@ -372,6 +372,9 @@
     <ClCompile Include="src\systems\DebugSystem.cpp">
       <Filter>Source Files\systems</Filter>
     </ClCompile>
+    <ClCompile Include="src\client\SendGetMethods.cpp">
+      <Filter>Source Files\client</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\stdafx.h">
